:recycle: api: move cpf usecases to session bounded context
