+++ api/src/shared/storage/infrastructure/providers/S3ObjectStorageProvider.js
