+++ api/tests/integration/infrastructure/jobs/cpf-export/handlers/create-and-upload_test.js
@@ -5,22 +5,24 @@ import { createUnzip } from 'node:zlib';
 
 import fs from 'fs';
 import lodash from 'lodash';
-const { noop } = lodash;
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc.js';
 import timezone from 'dayjs/plugin/timezone.js';
+import stream from 'stream';
+import * as url from 'url';
+
+const { noop } = lodash;
+
 dayjs.extend(utc);
 dayjs.extend(timezone);
 
-import stream from 'stream';
 const { PassThrough } = stream;
 
-import * as url from 'url';
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
 describe('Integration | Infrastructure | jobs | cpf-export | create-and-upload', function () {
   let cpfCertificationResultRepository;
-  let cpfExternalStorage;
+  let uploadCpfFiles;
   let clock;
   const expectedFileName = 'pix-cpf-export-20220102-114327.xml.gz';
   let logger;
@@ -51,7 +53,7 @@ describe('Integration | Infrastructure | jobs | cpf-export | create-and-upload',
       markCertificationCoursesAsExported: sinon.stub(),
     };
 
-    cpfExternalStorage = {
+    uploadCpfFiles = {
       upload: sinon.stub(),
     };
 
@@ -59,7 +61,7 @@ describe('Integration | Infrastructure | jobs | cpf-export | create-and-upload',
 
     uuidService.randomUUID.returns('xxx-yyy-zzz');
 
-    cpfExternalStorage.upload
+    uploadCpfFiles.upload
       .withArgs({
         filename: expectedFileName,
         readableStream: sinon.match(PassThrough),
@@ -76,7 +78,7 @@ describe('Integration | Infrastructure | jobs | cpf-export | create-and-upload',
       data: { batchId },
       cpfCertificationResultRepository,
       cpfCertificationXmlExportService,
-      cpfExternalStorage,
+      uploadCpfFiles,
       logger,
       uuidService,
     });
