+++ api/tests/unit/infrastructure/jobs/cpf-export/handlers/create-and-upload_test.js
@@ -2,21 +2,22 @@ import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 import { createAndUpload } from '../../../../../../lib/infrastructure/jobs/cpf-export/handlers/create-and-upload.js';
 import stream from 'stream';
 import * as uuidService from 'crypto';
-
-const { PassThrough, Readable } = stream;
-
 import lodash from 'lodash';
-const { noop } = lodash;
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc.js';
 import timezone from 'dayjs/plugin/timezone.js';
+
+const { PassThrough, Readable } = stream;
+
+const { noop } = lodash;
+
 dayjs.extend(utc);
 dayjs.extend(timezone);
 
 describe('Unit | Infrastructure | jobs | cpf-export | create-and-upload', function () {
   let cpfCertificationResultRepository;
   let cpfCertificationXmlExportService;
-  let cpfExternalStorage;
+  let uploadCpfFiles;
   let loggerSpy;
   let clock;
   let logger;
@@ -32,7 +33,7 @@ describe('Unit | Infrastructure | jobs | cpf-export | create-and-upload', functi
     cpfCertificationXmlExportService = {
       buildXmlExport: sinon.stub(),
     };
-    cpfExternalStorage = {
+    uploadCpfFiles = {
       upload: sinon.stub(),
     };
     logger = { error: noop, info: noop };
@@ -63,7 +64,7 @@ describe('Unit | Infrastructure | jobs | cpf-export | create-and-upload', functi
         data: { batchId },
         cpfCertificationResultRepository,
         cpfCertificationXmlExportService,
-        cpfExternalStorage,
+        uploadCpfFiles,
         logger,
       });
 
@@ -73,7 +74,7 @@ describe('Unit | Infrastructure | jobs | cpf-export | create-and-upload', functi
         writableStream: sinon.match(PassThrough),
         uuidService,
       });
-      expect(cpfExternalStorage.upload).to.have.been.calledWithExactly({
+      expect(uploadCpfFiles.upload).to.have.been.calledWithExactly({
         filename: 'pix-cpf-export-20220101-114327.xml.gz',
         readableStream: sinon.match(Readable),
       });
@@ -98,13 +99,13 @@ describe('Unit | Infrastructure | jobs | cpf-export | create-and-upload', functi
         data: { batchId },
         cpfCertificationResultRepository,
         cpfCertificationXmlExportService,
-        cpfExternalStorage,
+        uploadCpfFiles,
         logger,
       });
 
       // then
       expect(cpfCertificationXmlExportService.buildXmlExport).to.not.have.been.called;
-      expect(cpfExternalStorage.upload).to.not.have.been.called;
+      expect(uploadCpfFiles.upload).to.not.have.been.called;
       expect(cpfCertificationResultRepository.markCertificationCoursesAsExported).to.not.have.been.called;
 
       expect(loggerSpy).to.have.been.calledOnce;
